Move errors into internal folder
