+++ lib/templates/internal/errors.ts
