+++ lib/templates/updateFormBranding.ts
@@ -11,7 +11,7 @@ import {
 import { logMessage } from "@lib/logger";
 import { validateTemplate } from "@lib/utils/form-builder/validate";
 import { validateTemplateSize } from "@lib/utils/validateTemplateSize";
-import { InvalidFormConfigError } from "./errors";
+import { InvalidFormConfigError } from "./internal/errors";
 import { parseTemplate } from "./shared";
 
 /**
