+++ lib/templates/updateIsPublishedForTemplate.ts
@@ -4,7 +4,7 @@ import { FormRecord } from "@lib/types";
 import { authorization } from "../privileges";
 import { AuditLogAccessDeniedDetails, logEvent } from "../auditLogs";
 import { deleteDraftFormResponses } from "../vault";
-import { TemplateAlreadyPublishedError } from "./errors";
+import { TemplateAlreadyPublishedError } from "./internal/errors";
 import { getFullTemplateByID } from "./getFullTemplateByID";
 import { parseTemplate } from "./shared";
